Fix -Werror=format= failures with ESP-IDF v5.2.1

+++ main/ble_server.c
@@ -22,6 +22,7 @@
 #include "esp_err.h"
 #include "esp_task_wdt.h"
 #include "nvs_flash.h"
+#include <inttypes.h>
 #include "esp_bt.h"
 #include "string.h"
 
@@ -316,8 +317,8 @@ static void send_buffered_message(void)
     while(temp_spp_recv_data_node_p1 != NULL){
         /* Bounds check BEFORE memcpy to prevent buffer overflow */
         if(recv_len + temp_spp_recv_data_node_p1->len > sizeof(buf)) {
-            ESP_LOGW("ble_srv", "BLE recv buffer overflow prevented: %lu + %d > %u",
-                     (unsigned long)recv_len, temp_spp_recv_data_node_p1->len, (unsigned)sizeof(buf));
+            ESP_LOGW("ble_srv", "BLE recv buffer overflow prevented: %" PRIu32 " + %" PRId32 " > %u",
+                     recv_len, temp_spp_recv_data_node_p1->len, (unsigned)sizeof(buf));
             break;
         }
         memcpy(buf + recv_len, (char *)(temp_spp_recv_data_node_p1->node_buff), temp_spp_recv_data_node_p1->len);
